add dutch translation
As I'm not sure your app itself is/will be in Dutch, I used a combination of Settings/Instellingen instead of only the Dutch 'instellingen'. Feel free to adapt this as necessary. Also, do let me know if you need something else translated as well (such as labels in the app or something; again, I have no clue how this works haha).